massa_client

Module cmds

Structs

CommandIter

An iterator over the variants of Command

ExtendedWallet

Aggregation of the local, with some useful information as the balance, etc to be printed by the client.

ExtendedWalletEntry 🔒

Used to have a shinny json output TODO re-factor me

ListOperationIter

An iterator over the variants of ListOperation

Enums

Command

All the client commands the order they are defined is the order they are displayed in so be careful Maybe it would be worth renaming some of them for consistency Use props(pwd_not_needed = “true”) if the command does not need an access to the wallet, to avoid unnecessary Use props(hide_from_history = “true”) if the command contains sensitive information that should not be logged in the history file. prompting of the user.

ListOperation

Functions

convert_args_u256_to_u128 🔒

get_file_as_byte_vec 🔒

reads a file

help

Display the help of all commands

parse_key_value

parse_vec

TODO: ugly utilities functions takes a slice of string and makes it into a Vec<T>

send_operation 🔒

helper to wrap and send an operation with proper validity period