Enum massa_client::cmds::Command

pub enum Command {

    exit,
    help,
    node_unban_by_ip,
    node_unban_by_id,
    node_ban_by_ip,
    node_ban_by_id,
    node_stop,
    node_get_staking_addresses,
    node_remove_staking_addresses,
    node_add_staking_secret_keys,
    node_testnet_rewards_program_ownership_proof,
    node_whitelist,
    node_remove_from_whitelist,
    get_status,
    get_addresses,
    get_datastore_entry,
    get_block,
    get_endorsements,
    get_operations,
    get_filtered_sc_output_event,
    wallet_info,
    wallet_generate_secret_key,
    wallet_add_secret_keys,
    wallet_remove_addresses,
    wallet_sign,
    buy_rolls,
    sell_rolls,
    send_transaction,
    send_smart_contract,
    call_smart_contract,
    read_only_smart_contract,
    read_only_call,
    when_episode_ends,
    when_moon,

}

All the client commands the order they are defined is the order they are displayed in so be careful Maybe it would be worth renaming some of them for consistency

Variants

exit

help

node_unban_by_ip

node_unban_by_id

node_ban_by_ip

node_ban_by_id

node_stop

node_get_staking_addresses

node_remove_staking_addresses

node_add_staking_secret_keys

node_testnet_rewards_program_ownership_proof

node_whitelist

node_remove_from_whitelist

get_status

get_addresses

get_datastore_entry

get_block

get_endorsements

get_operations

get_filtered_sc_output_event

wallet_info

wallet_generate_secret_key

wallet_add_secret_keys

wallet_remove_addresses

wallet_sign

buy_rolls

sell_rolls

send_transaction

send_smart_contract

call_smart_contract

read_only_smart_contract

read_only_call

when_episode_ends

when_moon

Implementations

impl Command

pub(crate) fn help(&self)

Display the help of the command with fancy colors and so on

pub(crate) async fn run(
&self,
client: &Client,
wallet: &mut Wallet,
parameters: &[String],
json: bool
) -> Result<Box<dyn Output>>

run a given command

parameters

client: the RPC client

wallet: an access to the wallet

parameters: the parsed parameters

json: true if –json was passed as an option it means that we don’t want to print anything we just want the json output