Struct massa_client::rpc::RpcClient

pub struct RpcClient(TypedClient);

Tuple Fields
0: TypedClient
Implementations

impl RpcClient

Typed wrapper to API calls based on the method given by jsonrpc_core_client:

fn call_method<T: Serialize, R: DeserializeOwned>(
    method: &str,
    returns: &str,
    args: T,
) -> impl Future<Output = RpcResult<R>> {
}

pub(crate) async fn from_url(url: &str) -> RpcClient

Default constructor

pub(crate) async fn stop_node(&self) -> RpcResult<()>

Gracefully stop the node.

pub(crate) async fn node_sign_message(
&self,
message: Vec<u8>
) -> RpcResult<PubkeySig>

Sign message with node’s key. Returns the public key that signed the message and the signature.

pub(crate) async fn add_staking_private_keys(
&self,
private_keys: Vec<PrivateKey>
) -> RpcResult<()>

Add a vec of new private keys for the node to use to stake. No confirmation to expect.

pub(crate) async fn remove_staking_addresses(
&self,
addresses: Vec<Address>
) -> RpcResult<()>

Remove a vec of addresses used to stake. No confirmation to expect.

pub(crate) async fn get_staking_addresses(&self) -> RpcResult<AddressHashSet>

Return hashset of staking addresses.

pub(crate) async fn ban(&self, ips: Vec<IpAddr>) -> RpcResult<()>

Bans given node id No confirmation to expect.

pub(crate) async fn unban(&self, ips: Vec<IpAddr>) -> RpcResult<()>

Unbans given ip addr No confirmation to expect.

pub(crate) async fn get_status(&self) -> RpcResult<NodeStatus>

summary of the current state: time, last final blocks (hash, thread, slot, timestamp), clique count, connected nodes count

pub(crate) async fn _get_cliques(&self) -> RpcResult<Vec<Clique>>

pub(crate) async fn _get_stakers(&self) -> RpcResult<AddressHashMap<u64>>

Returns the active stakers and their roll counts for the current cycle.

pub(crate) async fn get_operations(
&self,
operation_ids: Vec<OperationId>
) -> RpcResult<Vec<OperationInfo>>

Returns operations information associated to a given list of operations’ IDs.

pub(crate) async fn get_endorsements(
&self,
endorsement_ids: Vec<EndorsementId>
) -> RpcResult<Vec<EndorsementInfo>>

pub(crate) async fn get_block(&self, block_id: BlockId) -> RpcResult<BlockInfo>

Get information on a block given its BlockId

pub(crate) async fn _get_graph_interval(
&self,
time_interval: TimeInterval
) -> RpcResult<Vec<BlockSummary>>

Get the block graph within the specified time interval. Optional parameters: from <time_start> (included) and to <time_end> (excluded) millisecond timestamp

pub(crate) async fn get_addresses(
&self,
addresses: Vec<Address>
) -> RpcResult<Vec<AddressInfo>>

pub(crate) async fn send_operations(
&self,
operations: Vec<Operation>
) -> RpcResult<Vec<OperationId>>

Adds operations to pool. Returns operations that were ok and sent to pool.