Module massa_client::cmds

Macros

client_warning 🔒

print a yellow warning

rpc_error 🔒

bail a shinny RPC error

Structs

CommandIter

An iterator over the variants of Self

ExtendedWallet

Aggregation of the local, with some useful information as the balance, etc to be printed by the client.

ExtendedWalletEntry 🔒

Used to have a shinny json output TODO re-factor me

ListOperationIter

An iterator over the variants of Self

Enums

Command

All the client commands the order they are defined is the order they are displayed in so be careful Maybe it would be worth renaming some of them for consistency

ListOperation

Functions

get_file_as_byte_vec 🔒

reads a file

help 🔒

Display the help of all commands

parse_key_value

parse_vec

TODO: ugly utilities functions takes a slice of string and makes it into a Vec<T>

send_operation 🔒

helper to wrap and send an operation with proper validity period