Enum massa_client::cmds::Command

pub enum Command {

    help,
    exit,
    node_unban_by_ip,
    node_unban_by_id,
    node_ban_by_ip,
    node_ban_by_id,
    node_stop,
    node_get_staking_addresses,
    node_start_staking,
    node_stop_staking,
    node_testnet_rewards_program_ownership_proof,
    node_bootstrap_whitelist,
    node_bootstrap_blacklist,
    node_peers_whitelist,
    get_status,
    get_addresses,
    get_datastore_entry,
    get_blocks,
    get_endorsements,
    get_operations,
    get_filtered_sc_output_event,
    wallet_info,
    wallet_get_public_key,
    wallet_get_secret_key,
    wallet_generate_secret_key,
    wallet_add_secret_keys,
    wallet_remove_addresses,
    wallet_sign,
    buy_rolls,
    sell_rolls,
    send_transaction,
    execute_smart_contract,
    call_smart_contract,
    read_only_execute_smart_contract,
    read_only_call,
    when_episode_ends,
    when_moon,

}

All the client commands the order they are defined is the order they are displayed in so be careful Maybe it would be worth renaming some of them for consistency Use props(pwd_not_needed = “true”) if the command does not need an access to the wallet, to avoid unnecessary prompting of the user.

Variants

help

exit

node_unban_by_ip

node_unban_by_id

node_ban_by_ip

node_ban_by_id

node_stop

node_get_staking_addresses

node_start_staking

node_stop_staking

node_testnet_rewards_program_ownership_proof

node_bootstrap_whitelist

node_bootstrap_blacklist

node_peers_whitelist

get_status

get_addresses

get_datastore_entry

get_blocks

get_endorsements

get_operations

get_filtered_sc_output_event

wallet_info

wallet_get_public_key

wallet_get_secret_key

wallet_generate_secret_key

wallet_add_secret_keys

wallet_remove_addresses

wallet_sign

buy_rolls

sell_rolls

send_transaction

execute_smart_contract

call_smart_contract

read_only_execute_smart_contract

read_only_call

when_episode_ends

when_moon

Implementations

impl Command

pub(crate) fn help(&self)

Display the help of the command with fancy colors and so on

pub(crate) fn is_pwd_needed(&self) -> bool

Returns true if the command needs wallet access

pub(crate) async fn run( &self, client: &mut Client, wallet_opt: &mut Option<Wallet>, parameters: &[String], json: bool ) -> Result<Box<dyn Output>>

run a given command

parameters

client: the RPC client

wallet_opt: an optional access to the wallet

parameters: the parsed parameters

json: true if –json was passed as an option it means that we don’t want to print anything we just want the json output